refactor: move isValidCssColor to utils
