+++ src/generic/pathway-card/PathwayCard.test.tsx
@@ -53,93 +53,6 @@ describe('PathwayCard', () => {
     expect(screen.queryByText('Bootcamp')).not.toBeInTheDocument();
   });
 
-  describe('CSS color validation', () => {
-    let cssDescriptor: PropertyDescriptor | undefined;
-    let documentDescriptor: PropertyDescriptor | undefined;
-
-    beforeEach(() => {
-      cssDescriptor = Object.getOwnPropertyDescriptor(globalThis, 'CSS');
-      documentDescriptor = Object.getOwnPropertyDescriptor(globalThis, 'document');
-    });
-
-    afterEach(() => {
-      if (cssDescriptor) {
-        Object.defineProperty(globalThis, 'CSS', cssDescriptor);
-      } else {
-        delete (globalThis as { CSS?: unknown }).CSS;
-      }
-      if (documentDescriptor) {
-        Object.defineProperty(globalThis, 'document', documentDescriptor);
-      } else {
-        delete (globalThis as { document?: unknown }).document;
-      }
-    });
-
-    it('returns false when document is unavailable', () => {
-      const originalDocument = documentDescriptor?.get?.call(globalThis) ?? documentDescriptor?.value;
-      let documentWasChecked = false;
-
-      Object.defineProperty(globalThis, 'CSS', {
-        configurable: true,
-        value: undefined,
-      });
-      Object.defineProperty(globalThis, 'document', {
-        configurable: true,
-        get: () => {
-          if (new Error().stack?.includes('isValidCssColor')) {
-            documentWasChecked = true;
-            return undefined;
-          }
-          return originalDocument;
-        },
-      });
-
-      render(<PathwayCard {...props} pathwayType="Bootcamp" pathwayTypeBackgroundColor="#123456" pathwayTypeTextColor="white" />);
-
-      expect(documentWasChecked).toBe(true);
-      expect(screen.getByText('Bootcamp').style.backgroundColor).toBe('');
-    });
-
-    it('uses CSS.supports when available and accepts valid colors', () => {
-      const supports = jest.fn().mockReturnValue(true);
-      Object.defineProperty(globalThis, 'CSS', {
-        configurable: true,
-        value: { supports },
-      });
-
-      render(<PathwayCard {...props} pathwayType="Bootcamp" pathwayTypeBackgroundColor="#123456" pathwayTypeTextColor="white" />);
-
-      expect(supports).toHaveBeenCalledWith('color', '#123456');
-      expect(supports).toHaveBeenCalledWith('color', 'white');
-      expect(screen.getByText('Bootcamp')).toHaveStyle({ backgroundColor: '#123456', color: 'white' });
-    });
-
-    it('uses CSS.supports when available and rejects invalid colors', () => {
-      const supports = jest.fn().mockReturnValue(false);
-      Object.defineProperty(globalThis, 'CSS', {
-        configurable: true,
-        value: { supports },
-      });
-
-      render(<PathwayCard {...props} pathwayType="Bootcamp" pathwayTypeBackgroundColor="not-a-color" pathwayTypeTextColor="white" />);
-
-      expect(supports).toHaveBeenCalledWith('color', 'not-a-color');
-      expect(screen.getByText('Bootcamp').style.backgroundColor).toBe('');
-    });
-
-    it('falls back safely when CSS.supports throws', () => {
-      const supports = jest.fn().mockImplementation(() => { throw new Error('unsupported'); });
-      Object.defineProperty(globalThis, 'CSS', {
-        configurable: true,
-        value: { supports },
-      });
-
-      render(<PathwayCard {...props} pathwayType="Bootcamp" pathwayTypeBackgroundColor="#123456" pathwayTypeTextColor="white" />);
-
-      expect(screen.getByText('Bootcamp')).not.toHaveStyle({ backgroundColor: '#123456' });
-    });
-  });
-
   it('uses default badge styling when either custom color is invalid or missing', () => {
     const { rerender } = render(
       <PathwayCard {...props} pathwayType="Bootcamp" pathwayTypeBackgroundColor="#123456" />,
